+++ client/src/scripts/others/constants.ts
@@ -1,9 +1,11 @@
+import { LevelDefinition } from "common/scripts/config/level_definition.ts";
+
 export type PlayArgs={
     type: "online"
     mode:string
     team_size:number
 }|{
     type: "campaign"
-    level:number
+    level:LevelDefinition
     dificulty:number
 }